// src/Presentation/gRPC/Stub/hotel_pb.ts
export interface EmptyResponse {
}
/**
 * @generated from protobuf message BookingRequest
 */
export interface BookingRequest {
    /**
     * @generated from protobuf field: string checkIn = 1;
     */
    checkIn: string;
    /**
     * @generated from protobuf field: string checkOut = 2;
     */
    checkOut: string;
    /**
     * @generated from protobuf field: int32 capacity = 3;
     */
    capacity: number;
}
/**
 * @generated from protobuf message BookingResponse
 */
export interface BookingResponse {
    /**
     * @generated from protobuf field: string bookingId = 1;
     */
    bookingId: string;
}
/**
 * @generated from protobuf message CancelBookRequest
 */
export interface CancelBookRequest {
    /**
     * @generated from protobuf field: string bookingId = 1;
     */
    bookingId: string;
}
/**
 * @generated from protobuf message CheckInRequest
 */
export interface CheckInRequest {
    /**
     * @generated from protobuf field: string bookingId = 1;
     */
    bookingId: string;
}
/**
 * @generated from protobuf message CheckOutRequest
 */
export interface CheckOutRequest {
    /**
     * @generated from protobuf field: string bookingId = 1;
     */
    bookingId: string;
}
/**
 * @generated from protobuf message SearchRoomRequest
 */
export interface SearchRoomRequest {
    /**
     * @generated from protobuf field: int32 capacity = 1;
     */
    capacity: number;
    /**
     * @generated from protobuf field: int32 floor = 2;
     */
    floor: number;
}
/**
 * @generated from protobuf message SearchRoomResponse
 */
export interface SearchRoomResponse {
    /**
     * @generated from protobuf field: repeated SearchRoomResponse.Room room = 1;
     */
    room: SearchRoomResponse_Room[];
}
/**
 * @generated from protobuf message SearchRoomResponse.Room
 */
export interface SearchRoomResponse_Room {
    /**
     * @generated from protobuf field: int32 id = 1;
     */
    id: number;
    /**
     * @generated from protobuf field: int32 capacity = 2;
     */
    capacity: number;
    /**
     * @generated from protobuf field: int32 floor = 3;
     */
    floor: number;
}
// @generated message type with reflection information, may provide speed optimized methods
class EmptyResponse$Type extends MessageType<EmptyResponse> {
    constructor() {
        super("EmptyResponse", []);
    }
}
/**
 * @generated MessageType for protobuf message EmptyResponse
 */
export const EmptyResponse = new EmptyResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class BookingRequest$Type extends MessageType<BookingRequest> {
    constructor() {
        super("BookingRequest", [
            { no: 1, name: "checkIn", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "checkOut", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "capacity", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message BookingRequest
 */
export const BookingRequest = new BookingRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class BookingResponse$Type extends MessageType<BookingResponse> {
    constructor() {
        super("BookingResponse", [
            { no: 1, name: "bookingId", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message BookingResponse
 */
export const BookingResponse = new BookingResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CancelBookRequest$Type extends MessageType<CancelBookRequest> {
    constructor() {
        super("CancelBookRequest", [
            { no: 1, name: "bookingId", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message CancelBookRequest
 */
export const CancelBookRequest = new CancelBookRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CheckInRequest$Type extends MessageType<CheckInRequest> {
    constructor() {
        super("CheckInRequest", [
            { no: 1, name: "bookingId", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message CheckInRequest
 */
export const CheckInRequest = new CheckInRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CheckOutRequest$Type extends MessageType<CheckOutRequest> {
    constructor() {
        super("CheckOutRequest", [
            { no: 1, name: "bookingId", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message CheckOutRequest
 */
export const CheckOutRequest = new CheckOutRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SearchRoomRequest$Type extends MessageType<SearchRoomRequest> {
    constructor() {
        super("SearchRoomRequest", [
            { no: 1, name: "capacity", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "floor", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message SearchRoomRequest
 */
export const SearchRoomRequest = new SearchRoomRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SearchRoomResponse$Type extends MessageType<SearchRoomResponse> {
    constructor() {
        super("SearchRoomResponse", [
            { no: 1, name: "room", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => SearchRoomResponse_Room }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message SearchRoomResponse
 */
export const SearchRoomResponse = new SearchRoomResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SearchRoomResponse_Room$Type extends MessageType<SearchRoomResponse_Room> {
    constructor() {
        super("SearchRoomResponse.Room", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "capacity", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "floor", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message SearchRoomResponse.Room
 */
export const SearchRoomResponse_Room = new SearchRoomResponse_Room$Type();